Complete re-write of Layout
