+++ Client/Components/PageHeader.razor
@@ -1,7 +1,7 @@
 ﻿<PageTitle>@PageEnum.Title</PageTitle>
 
-<div class="pb-1 mt-4 mb-2 border-bottom border-info">
-	<h2><i class="@PageEnum.Icon"></i> @PageEnum.Title</h2>
+<div class="pb-n1 mt-4 mb-3 border-bottom border-info">
+		<h2><i class="@PageEnum.Icon"></i>&nbsp;&nbsp;@PageEnum.Title</h2>
 </div>
 
 @code {
