+++ Client/Features/Home/MainBody.razor
@@ -1,0 +1,14 @@
+﻿<div class="card">
+  <div class="card-header">
+    Living Messiah Ministries
+  </div>
+  <div class="card-body">
+    <h5 class="card-title">Home Page Content</h5>
+    <p class="card-text">ToDo: Upcoming Events</p>
+    <p class="card-text">ToDo: Add the ShabbatMeter</p>
+  </div>
+</div>
+
+@code {
+
+}
