+++ Client/Features/Home/SidebarButtons.razor
@@ -1,0 +1,27 @@
+﻿@using Client.Enums
+
+<div class="list-group">
+	@foreach (var item in Nav.List.ToList()
+	.Where(w => w.Value == Nav.ShabbatService | w.Value == Nav.UpcomingEvents | w.Value == Nav.Donate | w.Value == Nav.Calendar)
+	.OrderBy(o => o.Sort))
+	{
+		<a href="@item.Index" title="@item.Title @HomeTitleSuffix"
+			 class="list-group-item list-group-item-warning ">
+			<div class="d-flex w-100 justify-content-between">
+				<i class="@item.Icon fa-fw fa-3x" aria-hidden="true"></i>
+				<span class="@TitleCSS">@item.Title</span>
+				<span class="float-end text-info">
+					<span class="hebrew">@HomeFloatRightHebrew</span>
+				</span>
+			</div>
+		</a>
+	}
+</div>
+
+@code {
+
+	protected string? HomeTitleSuffix = " Shalom  H7695";
+	protected string? HomeFloatRightHebrew = "שָׁלוֹם";
+	protected string? TitleCSS = "h6 ";
+
+}
