--- Client/Features/Home/TestNavSmartEnum.razor
@@ -1,85 +1,0 @@
-﻿@using Client.Enums
-
-<h2>Using <code>Nav.IsPartOfList</code></h2>
-
-<h4 class="mb-0">None</h4>
-	<ul class="list-inline ms-3">
-	@foreach (var item in Nav.List.ToList()
-		.Where(w => w.IsPartOfList(PageListType.None))
-		.OrderBy(o => o.Sort))
-	{
-		<li class="list-inline-item"> @item.Name [@item.Value], </li>
-	}
-</ul>
-
-<h4 class="mb-0">SitemapPage</h4>
-<ul class="list-inline ms-3">
-	@foreach (var item in Nav.List.ToList()
-		.Where(w => w.IsPartOfList(PageListType.SitemapPage))
-		.OrderBy(o => o.Sort))
-	{
-		<li class="list-inline-item"> @item.Name [@item.Value], </li>
-	}
-</ul>
-
-
-<h4 class="mb-0">Footer</h4>
-<ul class="list-inline ms-3">
-	@foreach (var item in Nav.List.ToList()
-	.Where(w => w.IsPartOfList(PageListType.Footer))
-	.OrderBy(o => o.Sort))
-	{
-		<li class="list-inline-item"> @item.Name [@item.Value], </li>
-	}
-</ul>
-
-<h4 class="mb-0">Layout</h4>
-<ul class="list-inline ms-3">
-	@foreach (var item in Nav.List.ToList()
-	.Where(w => w.IsPartOfList(PageListType.Layout))
-	.OrderBy(o => o.Sort))
-	{
-		<li class="list-inline-item"> @item.Name [@item.Value], </li>
-	}
-</ul>
-
-<h4 class="mb-0">Reply</h4>
-<ul class="list-inline ms-3">
-	@foreach (var item in Nav.List.ToList()
-	.Where(w => w.IsPartOfList(PageListType.Reply))
-	.OrderBy(o => o.Sort))
-	{
-		<li class="list-inline-item"> @item.Name [@item.Value], </li>
-	}
-</ul>
-
-<hr />
-
-<h2 class="mt-4 mb-0">Using <code>Nav.List</code></h2>
-<ul class="list-inline ms-3">
-	@foreach (var item in Nav.List.OrderBy(o => o.Sort))
-	{
-		<li class="list-inline-item"> @item.Name [@item.Value], </li>
-	}
-</ul>
-
-
-
-<h2 class="mt-4 mb-0">Using <code>Enum.GetValues(typeof(PageListType))</code></h2>
-<ul class="list-inline ms-3">
-	@foreach (int i in Enum.GetValues(typeof(PageListType)))
-	{
-		<li class="list-inline-item">@i, </li>
-	}
-</ul>
-
-<ul class="list-inline ms-3">
-	@foreach (var name in Enum.GetNames(typeof(PageListType)))
-	{
-		<li class="list-inline-item">@name, </li>
-	}
-</ul>
-
-	@code{
-
-	}
