--- Client/Layout/DELETE_MenuOffCanvas_and_NavMenu.txt
@@ -1,1 +1,0 @@
-when the menu is working satisfactorily these two components should not be needed any more
