+++ Client/Layout/MainLayout.razor
@@ -1,59 +1,77 @@
 ﻿@using Client.Layout.Navigation
+@using Client.Layout.Login
 @using Client.Enums
+
 @inherits LayoutComponentBase
 
-<div class="page">
-	<header>
-		<nav class="navbar bg-warning-subtle fixed-top">
+<div class="@MediaQuery.Xs.DivClass">
 
-			@* <MenuOffCanvas /> *@
-			@* <NavMenu /> *@
+	<div class="d-flex bg-warning-subtle">
+		<div class="p-1">
+			<HomeButton MediaQuery="@MediaQuery.Xs" />
+		</div>
 
-			<div class="container-fluid">
+		<div class="p-1 flex-grow-1">
+			<SearchForm />
+		</div>
+		<div class="">
+			<ToggleCollapseXs>
+				<AnchorButtons />
+			</ToggleCollapseXs>
+		</div>
+	</div>
 
-				<div class="@MediaQuery.SmOrMdOrLgOrXl.DivClass">
-					<HomeButton />
-				</div>
-
-				<SearchForm />
-				<button title="Navigation menu" class="navbar-toggler px-0 text-white" @onclick="ToggleNavMenu">
-					<span class="navbar-toggler-icon ms-0"></span>
-				</button>
-				<div class="@NavMenuCssClass nav-scrollable" @onclick="ToggleNavMenu">
-					<div class="offcanvas-body @NavMenuCssClass ">
-						<AnchorButtons />
-					</div>
-				</div>
-
-			</div>
-
-		</nav>
-
-	</header>
-
-	<main>
-		<article class="content mt-5 mb-2 px-1">
-			<LoginModal />
-			@Body
-		</article>
-	</main>
 </div>
+
+<div class="@MediaQuery.Sm.DivClass">
+	<div class="d-flex bg-warning-subtle">
+		<div class="p-1">
+			<HomeButton MediaQuery="@MediaQuery.Sm" />
+		</div>
+		<div class="p-1 flex-grow-1">
+			<SearchForm />
+		</div>
+		<div class="p-1">
+			<AnchorButtonsSmMdLgXl MediaQuery="@MediaQuery.Sm" />
+		</div>
+	</div>
+</div>
+
+<div class="@MediaQuery.MdOrLgOrXl.DivClass">
+	<div class="d-flex bg-warning-subtle">
+		<div class="p-1">
+			<HomeButton MediaQuery="@MediaQuery.MdOrLgOrXl" />
+		</div>
+		<div class="p-1 flex-grow-1">
+			<SearchForm />
+		</div>
+		<div class="p-1">
+			<AnchorButtonsSmMdLgXl MediaQuery="@MediaQuery.MdOrLgOrXl" />
+		</div>
+	</div>
+</div>
+
+<div class="d-flex justify-content-end mt-1 mx-1">
+	<ProfileButton />
+	<div class="p-2 fs-6">
+		<AnchorAndModal />
+	</div>
+</div>
+
+<div class="container-fluid">
+	@Body
+</div>
+
+
 <Footer />
-
-<div id="blazor-error-ui">
-	An unhandled error has occurred.
-	<a href="" class="reload">Reload</a>
-	<a class="dismiss">🗙</a>
-</div>
 
 
 @code {
+	protected string? cssClass = $"{Anchors.BtnSize.Xs} btn-info";
+		/*
 
-	private bool collapseNavMenu = true;
-	private string? NavMenuCssClass => collapseNavMenu ? "collapse" : null;
-	private void ToggleNavMenu()
-	{
-		collapseNavMenu = !collapseNavMenu;
-	}
+		.Where(w => w.Value == Nav.ShabbatService | w.Value == Nav.UpcomingEvents | w.Value == Nav.Donate | w.Value == Nav.Calendar)
 
-	}
+		*/
+
+}
