--- Client/Layout/NavMenu.razor
@@ -1,45 +1,0 @@
-﻿@using Client.Layout.Navigation
-
-<div class="top-row ps-3 navbar navbar-dark">
-	<div class="container-fluid">
-		
-		<HomeButton />
-		@* <a class="navbar-brand" href="">Client</a> *@
-
-		<button title="Navigation menu" class="navbar-toggler" @onclick="ToggleNavMenu">
-			<span class="navbar-toggler-icon"></span>
-		</button>
-	</div>
-</div>
-
-<div class="@NavMenuCssClass nav-scrollable" @onclick="ToggleNavMenu">
-	<nav class="flex-column">
-		<div class="nav-item px-3">
-			<NavLink class="nav-link" href="" Match="NavLinkMatch.All">
-				<span class="bi bi-house-door-fill-nav-menu" aria-hidden="true"></span> Home
-			</NavLink>
-		</div>
-		<div class="nav-item px-3">
-			<NavLink class="nav-link" href="counter">
-				<span class="bi bi-plus-square-fill-nav-menu" aria-hidden="true"></span> Counter
-			</NavLink>
-		</div>
-		<div class="nav-item px-3">
-			<NavLink class="nav-link" href="weather">
-				<span class="bi bi-list-nested-nav-menu" aria-hidden="true"></span> Weather
-			</NavLink>
-		</div>
-	</nav>
-</div>
-
-@code {
-
-	/* ToDo:	`NavMenu` IS NEVER REFERRENCED */
-
-	private bool collapseNavMenu = true;
-	private string? NavMenuCssClass => collapseNavMenu ? "collapse" : null;
-	private void ToggleNavMenu()
-	{
-		collapseNavMenu = !collapseNavMenu;
-	}
-}
