+++ Client/Layout/Navigation/AnchorButtonsSmMdLgXl.razor
@@ -1,0 +1,33 @@
+﻿@using Client.Enums
+
+@foreach (var item in Nav.List.ToList()
+		.Where(w => w.Value == Nav.Donate | w.Value == Nav.Calendar)
+		.OrderBy(o => o.Sort))
+{
+	<a href="@item.Index" title="@item.Title"
+		 type="button" class="@cssClass">
+		<i class="@item.Icon"></i>
+		@item.Title
+	</a>
+	<span>&nbsp;</span>
+}
+
+@code {
+	[Parameter] public string? ButtonColor { get; set; }
+	[Parameter] public string? ButtonSize { get; set; }
+	[Parameter, EditorRequired] public MediaQuery? MediaQuery { get; set; }
+
+	protected string? cssClass;
+
+	protected override void OnInitialized()
+	{
+		if (MediaQuery == MediaQuery.Sm)
+		{
+			cssClass = $"btn {ButtonColor ?? Anchors.BtnColors.Secondary} {ButtonSize ?? Anchors.BtnSize.Sm}";
+		}
+		else
+		{
+			cssClass = $"btn {ButtonColor ?? Anchors.BtnColors.Secondary} {ButtonSize ?? Anchors.BtnSize.Lg}";
+		}
+	}
+}
