--- Client/Layout/Navigation/DropDownAnchors.razor
@@ -1,23 +1,0 @@
-﻿@using Client.Enums
-
-<li class="nav-item dropdown">
-	<a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-		<i class="fas fa-ellipsis-h"></i> More
-	</a>
-
-
-	@*
-	ToDo: Differentiate in Nav if they are a dropdown menu item
-	foreach (var item in Nav.List.Where(w => w.IsDropDown).OrderBy(o => o.Value))
-	*@
-	<ul class="dropdown-menu">
-		<li>
-			<a class="nav-link" href="@Nav.Weather.Index"><i class="@Nav.Weather.Icon"></i> @Nav.Weather.Title</a>
-		</li>
-	</ul>
-
-</li>
-
-@code {
-
-}
