+++ Client/Layout/Navigation/HomeButton.razor
@@ -1,38 +1,63 @@
-﻿@using Page = Client.Enums.Nav
+﻿@using Client.Enums
+@using Page = Client.Enums.Nav
 
 <div class="d-sm-none">
-	<a class="navbar-brand nav-link" title="<576px Xs" href="@Page.Home.Index">
+	<a class="@cssClass" title="<576px Xs | @cssClass" href="@Page.Home.Index">
 		<i class="@Page.Home.Icon"></i>
 	</a>
 </div>
 
 
-@*Page.Home.Acronym*@
 <div class="d-none d-sm-block d-md-none d-lg-none d-xl-none">
-	<a class="navbar-brand nav-link " title="≥576px Sm" href="@Page.Home.Index">
+	<a class="@cssClass" title="≥576px Sm | @cssClass" href="@Page.Home.Index">
 		<i class="@Page.Home.Icon"></i>
 	</a>
 </div>
 
 <div class="d-none d-md-block d-lg-none d-xl-none">
-	<a class="navbar-brand nav-link " title="≥768px Md" href="@Page.Home.Index">
+	<a class="@cssClass " title="≥768px Md | @cssClass" href="@Page.Home.Index">
 		<i class="@Page.Home.Icon"></i>
 	</a>
 </div>
 
 <div class="d-none d-lg-block d-xl-none">
-	<a class="navbar-brand nav-link " title="≥992px Lg" href="@Page.Home.Index">
+	<a class="@cssClass " title="≥992px Lg | @cssClass" href="@Page.Home.Index">
 		<i class="@Page.Home.Icon"></i>
 	</a>
 </div>
 
 <div class="d-none d-xl-block">
-	<a class="navbar-brand nav-link " title="≥1200px Xl" href="@Page.Home.Index">
+	<a class="@cssClass " title="≥1200px Xl | @cssClass" href="@Page.Home.Index">
 		<i class="@Page.Home.Icon"></i>
 	</a>
 </div>
 
 
 @code {
+	[Parameter]	public string? ButtonColor { get; set; } 
+	[Parameter] public string? ButtonSize { get; set; } 
+	[Parameter, EditorRequired] public MediaQuery? MediaQuery { get; set; }
+
+	protected string? cssClass;
 	
+	protected override void OnInitialized()
+	{
+		if (MediaQuery == MediaQuery.Xs | MediaQuery == MediaQuery.Sm)
+		{
+			cssClass = $"btn {ButtonColor ?? Anchors.BtnColors.Secondary} {ButtonSize ?? Anchors.BtnSize.Sm}";
+		}
+		else
+		{
+			cssClass = $"btn {ButtonColor ?? Anchors.BtnColors.Secondary} {ButtonSize ?? Anchors.BtnSize.Lg}";
+		}
+	}
+
+	/*
+		ToDo: what's special about this ... `navbar-brand nav-link`
+
+		[Parameter] public string? Title { get; set; } // = Meta.Acronym  Page.Home.Acronym
+		protected string? title = Meta.Acronym;
+	*/
+
+	// 
 }
