--- Client/Layout/Navigation/LoginDisplay_REVIEW.txt
@@ -1,79 +1,0 @@
-
-
-# LoginAnchor
-@* <a href="@Page.Login" class="btn btn-xs btn-outline-info float-end text-black-50">
-	<u>@Client.Enums.NavAccount.Login.Title</u> <i class="@Client.Enums.NavAccount.Login.Icon"></i>
-</a>
- *@
-@code {
-
-}
-
-
-# LoginDisplay
-@using static LivingMessiah.Web.Links.Account
-@using LivingMessiah.Web.Shared
-@using LivingMessiah.Web.Shared.Header.Enums
-@using LivingMessiah.Web.Enums
-
-<div class="container-fluid bg-light">
-
-	<div class="@MediaQuery.XsOrSm.DivClass">
-
-		<div class="row pt-2 pb-1 d-print-none">
-			<div class="col-12">
-				<AuthorizeView>
-					<Authorized>
-						<LogoutForm />
-						<ProfileAnchor IsXsOrSm="true" />
-					</Authorized>
-
-					<NotAuthorized>
-						<div class="col-12">
-							<LoginAnchor />
-						</div>
-					</NotAuthorized>
-				</AuthorizeView>
-			</div>
-		</div>
-
-		<div class="row pt-2 pb-1 d-print-none">
-			<div class="col-12">
-				<BibleSearch />
-			</div>
-		</div>
-
-	</div>
-
-	<div class="@MediaQuery.MdOrLgOrXl.DivClass">
-
-		<div class="row pt-2 pb-1 d-print-none">
-			<div class="col-6">
-				<BibleSearch />
-			</div>
-			<div class="col-6">
-				<AuthorizeView>
-					<Authorized>
-						<LogoutForm />
-						<ProfileAnchor IsXsOrSm="false" />
-					</Authorized>
-
-					<NotAuthorized>
-						<div class="col-12">
-							<LoginAnchor />
-						</div>
-					</NotAuthorized>
-				</AuthorizeView>
-
-			</div>
-		</div>
-	</div>
-
-	<div class="row pt-2 pb-1 d-print-none">
-		<div class="col-12">
-			<BookChapterAnchorList />
-		</div>
-	</div>
-
-</div>
-
