+++ Client/Layout/Navigation/SearchForm.razor
@@ -1,24 +1,24 @@
 ﻿@using Blazored.Typeahead
 @using Client.Enums
 
-<div class="d-flex" role="search">
-	<BlazoredTypeahead SearchMethod="SearchNavs"
-										 TValue="Nav"
-										 TItem="Nav"
-										 Value="CurrentNav"
-										 ValueChanged="SelectedResultChanged"
-										 ValueExpression="@(() => CurrentNav)"
-										 EnableDropDown="true"
-										 MaximumSuggestions="40"
-										 MinimumLength="2"   
-										 placeholder="Search">
-		<SelectedTemplate Context="mycontext">
-			@mycontext!.Title
-		</SelectedTemplate>
-		<HelpTemplate>2+ characters</HelpTemplate>
-		<ResultTemplate Context="mycontext">
-			<i class="@mycontext!.Icon"></i> @mycontext.Title
-		</ResultTemplate>
-	</BlazoredTypeahead>
+@* <div class="d-flex" role="search"></div> *@
 
-</div>
+<BlazoredTypeahead SearchMethod="SearchNavs"
+									 TValue="Nav"
+									 TItem="Nav"
+									 Value="CurrentNav"
+									 ValueChanged="SelectedResultChanged"
+									 ValueExpression="@(() => CurrentNav)"
+									 EnableDropDown="true"
+									 MaximumSuggestions="40"
+									 MinimumLength="2"
+									 placeholder="Search">
+	<SelectedTemplate Context="mycontext">
+		@mycontext!.Title
+	</SelectedTemplate>
+	<HelpTemplate>2+ characters</HelpTemplate>
+	<ResultTemplate Context="mycontext">
+		<i class="@mycontext!.Icon"></i>&nbsp;&nbsp;@mycontext.Title
+	</ResultTemplate>
+</BlazoredTypeahead>
+
